requirements.txt
functions-framework==3.*
requests==2.31.0
flask==3.0.*
python-dotenv==1.0.1